// src/api/models/DsMessageWithUsers.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Mailing app
 * my api
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DsMessageWithUsers
 */
export interface DsMessageWithUsers {
    /**
     * 
     * @type {string}
     * @memberof DsMessageWithUsers
     */
    creator?: string;
    /**
     * 
     * @type {string}
     * @memberof DsMessageWithUsers
     */
    dateCreate?: string;
    /**
     * 
     * @type {string}
     * @memberof DsMessageWithUsers
     */
    dateFinish?: string;
    /**
     * 
     * @type {string}
     * @memberof DsMessageWithUsers
     */
    dateUpdate?: string;
    /**
     * 
     * @type {number}
     * @memberof DsMessageWithUsers
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof DsMessageWithUsers
     */
    moderator?: string;
    /**
     * 
     * @type {string}
     * @memberof DsMessageWithUsers
     */
    status?: string;
    /**
     * 
     * @type {string}
     * @memberof DsMessageWithUsers
     */
    text?: string;
}

/**
 * Check if a given object implements the DsMessageWithUsers interface.
 */
export function instanceOfDsMessageWithUsers(value: object): value is DsMessageWithUsers {
    return true;
}

export function DsMessageWithUsersFromJSON(json: any): DsMessageWithUsers {
    return DsMessageWithUsersFromJSONTyped(json, false);
}

export function DsMessageWithUsersFromJSONTyped(json: any, ignoreDiscriminator: boolean): DsMessageWithUsers {
    if (json == null) {
        return json;
    }
    return {
        
        'creator': json['creator'] == null ? undefined : json['creator'],
        'dateCreate': json['date_create'] == null ? undefined : json['date_create'],
        'dateFinish': json['date_finish'] == null ? undefined : json['date_finish'],
        'dateUpdate': json['date_update'] == null ? undefined : json['date_update'],
        'id': json['id'] == null ? undefined : json['id'],
        'moderator': json['moderator'] == null ? undefined : json['moderator'],
        'status': json['status'] == null ? undefined : json['status'],
        'text': json['text'] == null ? undefined : json['text'],
    };
}

export function DsMessageWithUsersToJSON(json: any): DsMessageWithUsers {
    return DsMessageWithUsersToJSONTyped(json, false);
}

export function DsMessageWithUsersToJSONTyped(value?: DsMessageWithUsers | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'creator': value['creator'],
        'date_create': value['dateCreate'],
        'date_finish': value['dateFinish'],
        'date_update': value['dateUpdate'],
        'id': value['id'],
        'moderator': value['moderator'],
        'status': value['status'],
        'text': value['text'],
    };
}